Refactor project packages
