+++ pom.xml
@@ -70,7 +70,6 @@
             <groupId>org.jenkins-ci.main</groupId>
             <artifactId>maven-plugin</artifactId>
             <version>2.0</version>
-	    <scope>provided</scope>
         </dependency>
         <!-- Tests -->
         <dependency>
