+++ src/main/resources/org/axway/grapes/jenkins/AdministrateResendAction/manage.jelly
