+++ src/main/resources/org/axway/grapes/jenkins/AdministrateResendAction/message.jelly
