+++ src/main/resources/org/axway/grapes/jenkins/GrapesNotifier/global.jelly
